Done a thing (really badly)
So you can now choose and enemy... but the way I did it was to have 2
lists, one arraylist to loop over and print each entry, and a HashMap of
the same things, so you can choose an entry based on user input. A
better way would be to number each enemy according to it's number in the
arraylist, and then take in a number as the user input, convert it from
a string to an int, and then get the enemy associated with that number
from the arraylist.