10.1
FIXED:
-PC now dies
BROKEN:
-You can't restart when you die in combat. This is because the if
statement in combat() returns, stopping the method from running, however
ProcessRoom does not because as far as it is concerned the player is not
dead, so it waits for player input instead of moving the player to
firstRoom.
-When it is your turn in combat, it is possible to type in any number,
so typing in a number other than those listed will cause the game to
crash.
-It is also possible to attack NPCs not in your current room by typing
in their number.
ADDED:
-You can now try and punch down locked doors, each time your fail you
take some damage. May one day give doors health, but I don't feel like
it's really necessary right now
-New Window class, to draw stuff in a window