10.3.1
When you are in the same room as a friendly npc, and you type in attack,
all the npcs in the room become enemies, and combat begins. Only it
doesn't, because when you type in attack, none of the stuff in the if
statement gets done. It just goes on to the next if. I DON'T UNDERSTAND