10.4.1
Am trying to change the code to use scenes and the new window thing I've
done instead of the console. Is harder than I originally thought. When a
decision is possible, I want buttons for each choice (e.g. north, south,
east, west). And instead of printing to the console, I want to change
the text for the scene/add text.
ADDED:
-new method to make scenes.
-same for buttons
-have begun implementing scenes