more stuff
Display (frontend) and Game (backend) now have the same instance of
Delegator

in order to use Move, 8 different inputs need to be able to call it
(north, northEast, East, etc). Have tried to add them to the
actionHandlers Map, but the syntax seems to be different to that of a
HashMap.