Fixing nullpointers
So the instance of Game which ActionHandler had was null, so I made it
so that when an actionHandler is created you need to give it an instance
of Game. However, it is Delegator which makes them, and it does not have
Game. Unfortunately, Game needs Display in it's constructor, which needs
Delegator in it's constructor.