Another update
When you moved, it would print the text for the room you just left,
rather than the one you went into. That is now fixed, and it now prints
the first room's text upon the game starting (there is no other way to
trigger it without moving). Working on a way to tell the user which
directions they can move in. Currently it just prints out the list,
which looks really ugly.