RollActions!! Woot!
I can't just give Display Game's list of traps when it is defined,
because Game uses Display when it's defined, AND I'M NOT GOING THROUGH
THAT AGAIN.

Instead, I made a method in Display, which recieves an ArrayList of
Traps, and in Main I give it Game's ArrayList of Traps. Game originally
only had a HashMap of Traps, so I had to make an ArrayList of them as
well, because I don't know how to loop through a HashMap (I did it ages
ago, but it didn't work).

Now I've got a if function that checks if the input is == roll, and if
so, loops through the list of traps to find the one that triggered, and
then sends a RollAction to Delegator using the Trap.

Delegator takes a String, and a list of parameters. the String is used
to identify the actionHandler to use, and I haven't worked out what the
parameters are used for yet.
I need to make a new ActionHandler maker for RollActions.