More RollActions! What a suprise!
This doesn't work. The idea was to check which trap you a rolling away
from, but makeRollAction is used when Game is initialised, so that
doesn't work.

Instead, I need to just use the one set of success/failure text things
for each type of roll, and have the user type in why they're rolling