Game no longer ends whenever you roll, regardless wether you succeeded or not.
I've also made it so that you can't leave the room until you've rolled
for the trap. I tried making it so that the room doesn't print if it has
a trap, but then i realised that you'd have to reenter the room to get
the directional text, so I got rid of that bit. I'm sure there is a way
to make it work, but idk what it is.