RollAction really needs to know which trap you're rolling for
So i thought it didn't really matter, but it does because otherwise you
are never able to leave a room with a trap in it
Need a global trap variable that MoveAction makes equal to the trap it
has found, and then RoomAction takes that variable, and makes the trap's
hasSprung true (if you've rolled, then it has sprung - you're either
dead or you dodged it)

THATS NOT TRUE IM SO DUMB.
All I had to do was give RollAction Game when it is created, then it can
make game.getCurrentRoom.hasSprung = true.
It all works now! YAY

also traps might be done now