Worked out why ItemActions weren't working
Turns out it was an old problem that I had no idea existed and was
nothing to do with any new code.

Made it so that when you go into a room it prints the item text, need to
add another string to Items though so that they have a string for when
you find them and one for when you pick them up.
^have done that now