Trying to add Buttons
It can't be this hard.

I copied what someone else on the internet did, not sure which part I
needed to do but it works now.

Unfortunately it isn't possible to add a List of buttons to the JPanel,
and so I'd have to use a method to create and add them to the panel, but
then I wouldnt be able to turn them on or off as I wouldn't have access
to them, and they wouldn't have names. So that wouldn't work.
I'm sure theres a better way to do this but for now I'm just going to
have to add each one line by line.

Now I need to be able to turn the buttons on or off according to which
directions the current room has. It is Display that has the window, so
Display needs to tell Window which buttons to have enabled/disabled. If
Window was in Game like it used to be it would be much easier but thatd
be dumb. How can display know which directions are avalible? for now
i'll just not disable any and just have the window display some text
saying you can't go that way.

Am struggling to make the text printed in display appear in the Window.
What I need to do is have an HashMap of the buttons, with the key being
the same string as the direction in room. Then I need to loop through
all the directions and change the button to be enabled/disabled
accordingly. I seem to remember having difficulty looping through
HashMaps but I can probably work it out? Maybe?
Actually no what I need to do is modify Outcome so that it contains some
way for MoveAction to send the possible directions to Display, which
gives them to Window.
Still not sure how to display the string properly in Window though

I also need to display the text for the room. I need to have multiple
strings displayed at once because otherwise traps and items wouldn't
work very well. I could use an ArrayList, and add Strings to it, and
just display the entire list.