Window stuff
Thought it wouldnt be possible to give Window the directions because I'm
returning an ArrayList of outcomes, rather than just the one, then
realised that I can loop through all the outcomes and check if their
directions ArrayList is null.
Also realised that I can give the x,y,width,height to makeButton and set
the button size and everything in there instead.

It should work, not sure why it doesn't. I've given the window a list of
keys which are identical to the keys for the buttons, but it's not
changing the buttons from enabled to disabled or vice versa. I tried
debugging it but I coulnd't see what was going wrong. :<

Worked it out, it's because I'm only looing throught the list when
Window is first initialised, and never again afterwards.

Found another problem, disabling buttons if the key doesn't get a
buttons obviously doesn't work. Haven't got time to fix that yet.