The buttons now change!!!!!!!!!!!!!!!1!!!!1!!!!!!!!!!!!!!!one!!!!!!
If directions has stuff in it, then all the buttons are set to be
disabled, and then I loop through all the buttons and enable the ones
for the current room. For some reason northwest,southeast, etc don't
work because I capitalised the second word in Window.
Also I've got the text to show in the top middle of the window. Had to
steal some code off the internet to get the length of the text in
pixels, so I can centre it. However it is only the room text, need to
give the Window an ArrayList of Strings so I can have room text, trap
text, and item text displayed at the same time.