Buttons actually do work
I was wrong, buttons do actually work. I just wasn't printing any text
from the actions. However they don't turn themselves on/off anymore, so
I need to fix that.

Fixed.

buttons now change, actionPerformed now calls update() each time an
action happens, so it can turn buttons on and off.

Now I need a button to roll for traps, only active when there is a trap.
However, that means I need to change how actionPerformed works so that
it doesn't think that "trap" is a direction and tries to delegate it as
a moveAction.

That means that actionPerformed needs a way to tell what kind of action
to delegate the button press as. I can just check if the key == "trap"
and if it is, delegate rollAction, if not, delegate moveAction.
That's done now, but I need to be able to activate the Roll button when
a room has an unsprung trap. Thats something I need to do in MoveAction,
as thats where I do traps n stuff. Does that mean I need to add another
variable to Outcome though? I need a way to tell Display that there's a
trap.

Done that, but it doesn't work yet. However, I've found another problem
- you're no longer locked to a room with a trap in it, you can just
leave the room, although it does print some text telling you off.
Not sure how to fix that. Not even sure why it's happening.