buttonssss
So the button to roll doesn't become enabled ever, but it should. In
moveAction I make the outcome.isTrap true, then in actionPerformed it
isn't true anymore? tf?
Actually no it turns on when you try to leave the room. that needs
fixing.
Fixed, but the button doesn't get enabled despite getting enabled.