SAVED
I fixed it!
In update(), it checks to see which directions are avalible, and
disables all the buttons that aren't in directions, which roll obviously
never is, and so gets deactivated every loop. Is now fixed.

Started actually doing stuff with rolls now. You can leave a room
regardless what you roll currently, so I need to make it so that you die
when you fail.