WHY WON'T YOU DIE
killing the player is harder than I thought it would be.

I can disable all buttons, but I can't actually end the game. Nevermind,
just means I can add a restart button that appears when u ded
Unfortunately it only updates the buttons when move into another room,
which is annoying. Tried a few things but I can't seem to fix that.

Trying to fix the room messages, which keep overlapping each other. I
worked out how to remove everything from the panel, which means I have
to re add all the buttons each loop, but it displays the text for the
room after you leave it. not sure why. Actually that might be a problem
I've had for ages but I've just not noticed it before.
Moving the code that removes the labels doesn't help, just means that
text never appears at all.