Text Fixes
Why would the text only appear after I've left the room? I don't get it.
debugging is unhelpful, as I can see that it is recieving the text to
display after you leave the room, but why doesn't it get it at the same
time as the console? Actually come to think of it they should be in the
same place.

Fixed.

All I had to do was put the bit that sets the text into display()
instead of it being in update().
Have also now made it so that text shows when the game first starts, and
removed the text telling you which directions you can go in.

Guess I can add ItemActions now.
Also realised I don't need a new arrayList for each type of button.