How to not leave room?
When you enter a room with a trap in it, no text displays, roll button
activates. Then if you click a directional button to leave the room, you
leave it, but then are unable to move again until you click the roll
button, which is deactivated because you're not longer in a room with a
trap.

I can fix it by disabling all buttons except the roll button when you
enter a room with a trap, which would make sense to do anyway. However I
don't know why the text doesn't display. It used to, what did I break?

Some part of the code isn't seeing the trap at the right time. Part of
it is though because otherwise the roll button wouldn't activate.
Have somewhat fixed it by moving the check to see if the room is
leaveable.