I thought I hadn't fixed it last night
apparently I did, all works fine now.
I need to make text for picking up items appear. Text appears in the
console, but not in the window. In fact all the text in the window
disappears when I take an item.
I debugged it and the text should be showing, but it just isn't there?
I haven't fixed it but Im done with coding for today and I want to save
my notes