More stuff
I can't work out why text won't display when i pick up an item, so I'm
gonna move on and get Liam to help at some point.
What next tho?
Combat?
Yeah why not. That'll break everything all over again.

So combat won't be just one action. When I enter combat, the entire
window needs to change. Seperate JPanel just for combat. Different
buttons for different actions - attack, flee, etc.
When you enter a room which has a enemy (that isn't dead), you enter
combat, panel changes, and initiative must be rolled. Then all
entities(?) follow the order of initiative. I have a system already
(commented out) that does some of this, I just need to implement it
again.

How do I start combat? It used to be a method inside Game(), I could
leave it as that, and call it? Probably. I'll try just leaving it and if
it turns out it needs to be a seperate class I'll do that.

How to send strings to Display()? Combat will be started by
MoveAction(), but I don't want to send the Strings via MoveAction() to
Display(). I could do that, but it feels a bit roundabout.