Actually doing combat stuff now
Firstly I need to make it so that the room isn't leaveable. Then I need
to make it so that the combat text ACTUALLY PRINTS WHEN YOU ENTER THE
ROOM NOT WHEN YOU LEAVE IT WHY DOES THAT EVEN HAPPEN.

I think I messed up somewhere because the isRoomLeaveable boolean seems
to now prevent me from entering the room in the first place. Or any
rooms at all in fact.
It seems to be acting like there's an enemy in the first room, but there
isn't.

Ok so that happens because I'm checking if the room's enemy list is !=
null, and every room's enemylist is !=null. Need to change that to
enemyList.get(0) != null or something like that.

BAM DONE