Combat Panel
Starting to make a new panel for combat stuff.

Panel made, I've also made methods that do all the panel stuff that
would be done when Display is first made, except I want to be able to
switch between panels, so I need methods that do all that for me,
because there's quite a bit to do.

made a boolean for combat, use that for combat stuff k bye