I RESET IT
IM SO DUMB HOW DID IT TAKE SO LONG FOR ME TO WORK OUT HOW TO DO THIS.

I've started doing the CardLayout thing again, and I'm not sure how/if I
can make it work. It seems to rely on ItemEvents, which I'm not using.

I've kinda got it working. However combatPanel seems to be empty?