Combat works to an extent
Combat works, but I'm calling it before I am actually in a room with
enemies, so it doesn't do anything. That means I can't call it from
MoveAction, as it needs to be call when the currentRoom has enemies in
it, and MoveAction will only detect that when the player tries to leave
the room.

I can't change MoveAction to help with that, as it is only called when I
click a button to leave a room. Instead I need a check when I enter a
room? No, when the next room has an enemy in it, I want to start combat
once I've entered it.

Maybe I do actually need a CombatAction. It seems a bit wrong to have
combat stuff like this in MoveAction.

ALRIGHT IM DOIN IT BOIS.

When do I use this though? It needs to be called if MoveAction says that
combat = true; But I check that when I'm looping through outcomes(), and
it'll break if I try to change the ArrayList I'm looping through. I
could call it from MoveAction I guess? Would that work though. I can't
really see any other way I can do it tbh. Maybe I can put it in
update().

I THINK THAT WORKED!
More tomorrow. Need to make the attack and flee buttons in combatPanel
actually do stuff.