Buttonssssssssssss
I think i've used that title before

Attack button. I need to be able to interact with combat() in game via
buttons in Display. I can't do that currently, because I only do stuff
with combat() once it has returned, and I need to do things while it is
running. At certain points in combat() I need to ask displa() for a
button press, and based on that press, do a roll.

Game and Display aren't supposed to interact directly. So how am I gonna
do this?

New Action, when waiting for the player to do something combat() is
waiting for a NewAction (idk what to call it, ive already got Action and
CombatAction) and when a NewAction is performed it changes something in
Game(). combat() is looping waiting for one of two variables to change,
and when you click the attack button, the attack boolean gets made=true.
How do loop until something is true?

Its ok i found it on the internets.

Ive begun to make the try;catch work but its not done yet