Having a combat loop is a bad idea and Ima use Actions instead
I need a few different Actions, instead of a loop. I need a loop of some
description though, because otherwise Initiative won't work.

Game needs to have a list of the entities currently in combat (and not
dead) And then on each one's turn, do a combat Action, then move to the
next one? So it gets the first item in the list, if it is an agressive
NPC, attackAcion on the PC, if it is the PC, then get input from Display
on whether to do an attackAction/fleeAction.