Hmm
akshdjabd ok so DoCombatAction. Recieves  the Character at the top of
the list, then if it is an NPC, it attacks the PC. First problem is that
AttackAction() requires a target and an attacker, not just the attacker.
Can I change that? Not really, the idea is that DoCombatAction is what
decides what to do, not AttackAction, which is supposed to do it. So
therefore DoCombatAction needs to have game.playercharacter, and the
NPC. HOWEVER, what if it is the PC's turn? then we need an NPC to
attack, provided the player decides to attack, so I need to get the
first NPC in the List.

In Game.combat(), get the character at the top of the list, then if it
is an NPC, send NPC and PC to DoCombatAction. If the character is at the
top of the list then send PC, and the first agressive NPC in the list to
DoComAct.

Display calls BeginCombatAction, which starts Game.combat(), and then
game.combat() calls DoCombatAction