How does Delegator work?
Display and Game both have the same Delegator. For some reason I'm
adding all the ActionHandlers to the Delegator in Game() but it works so
I assume thats not going to screw me over massively later on
ahhahahahahaha. I would just have to make the actionhandler in Display()
and then call it from DoCombatAction. DoCombatAction will need the same
Delegator though.