Welp
Im not sure how I can do this tbh

I need to know when the player has clicked a button. Looping is a bad
idea. I could have DoCombatAction ask a new class for button clicks each
loop, and display give that class a button click. That sounds workable,
but I dont want to use loops if I can avoid it. Maybe I could have
display send an action to DoComAction to tell it when its clicked. that
makes more sense ima do that.

I am not sure if I can actually do this without a loop, as I need to
continually check if actionType is null, and I am pretty sure it will be
initially. I may have to ask for help OH NO