Sod it
Im trying doing a loop, if it doesnt work, then never mind.
What I have now:
-When pc enters a room with an enemy in it, Display tells Game to begin
combat using BeginCombatAction.
-BeginCombatAction gets the enemies from the currentRoom and gives them
to game.combat().
-game.combat() does the initiative stuff, and then calls DoCombatAction
with the appropriate parameters.
-If it is the PC's tun, DoCombatAction waits for Display to give it
either "attack" or "flee".
-Then DoCombatAction delegates accordingly.

TODO:
It doesn't make sense if Display has to tell Game that combat needs to
start. Game shouldn't have to tell Display that the room has enemies,
and then have Display tell Game that. Game should start combat as soon
as it enters the room. The idea is that Game tells Display what to do,
not the other way around.
Need to check and see if this actually works.
Make a FleeAction (partially done)
Change Outcome so that it has a String variable that I can use for
various things, instead of a bunch of booleans.