Summary
Changed Outcome to have an ArrayList of Strings instead of just the one,
because the one outcome might need to have more than one variable.

Flee/AttackAction return stuff to DoAction, which now needs to do
something with that, If it is a sucessful FleeAction, then I need to
call a MoveAction that sets the player back one room, if it is a
sucessful AttackAction, then I need to tell Game that one of the enemies
is dead.

How do this:
DoAction gets the list of outcomes when it delegates the action, so it
needs to know if it was sucessful or not (which there is already a
variable for) and then know what to do with it. DoAction needs to send
the text to Display to be delegated, and any extra information to Game.
So DoAction needs to know which actionType it is so it knows what to do
with the results.