Good job I checked
There was a bit I forgot about, which is PlayerAction, writing to
actionType. Currently PlayerAction is never called, which means that
actionType will never have a value. That needs to be changed before I do
anything else.

Started work on that, but I need to finish it off in Display @ line 226