Updoot
Added ifs for attack and flee actions, and I should be able to call
PlayerAction from Display now.
Now I need to make it so that fleeAction calls a MoveAction to move you
to the previous room. To do that, I need to know which room was the last
room. Currently I don't think I have any way of getting that as I've
only needed currentRoom and nextRoom thus far. Therefore I need to have
a variable in Game called previousRoom, which MoveAction sets equal to
currentRoom immediately before making currentRoom equal to nextRoom.

Apparently I've already got all that. Im so clever