baaaack
Game.combat:
This doesn't make any sense to me, Combat starts when player enters a
"combat room" with hostile creatures
* then on player's turn they pick either fight or flight. Display then
sends this information to this function, which then decides what to do
with it.
* Having a "doCombat" does make sense, but why is it being called
immediately? It should be called in response to something.
*
* Surely PlayerAction should know what action the player has taken, and
then delegate that itself? Although we still then have the problem of
communicating with this function.