updoot
I need to make sure that doCombatAction isn't null when we start combat.
Currently when the player enters a combat room, when they press a
button, PlayerAction tries to give it to doCombat, but it is null, so
game breaks. Currently it is Game that makes the doCombatAction, but it
is possible that the combat function in Game isn't being called

Yeah thats the problem, BeginCombatAction is never getting made/called
so therefore doCombat can't be made/used because it is beginCombat that
does all that.

Combat in Game needs to make a BeginCombatAction when it starts